+++ packages/svelte/src/internal/server/index.js
@@ -4,9 +4,8 @@ import {
 	UNINITIALIZED,
 	DOMBooleanAttributes,
 	RawTextElements,
-	disallowed_paragraph_contents,
-	interactive_elements,
-	is_tag_valid_with_parent
+	ELEMENT_PRESERVE_ATTRIBUTE_CASE,
+	ELEMENT_IS_NAMESPACED
 } from '../../constants.js';
 import { escape_html } from '../../escaping.js';
 import { DEV } from 'esm-env';
@@ -16,28 +15,9 @@ import { validate_store } from '../shared/validate.js';
 
 /**
  * @typedef {{
- * 	tag: string;
- * 	parent: null | Element;
- * }} Element
- */
-
-/**
- * @typedef {{
  * 	head: string;
  * 	html: string;
  * }} RenderOutput
- */
-
-/**
- * @typedef {{
- * 	out: string;
- * 	anchor: number;
- * 	head: {
- * 		title: string;
- * 		out: string;
- * 		anchor: number;
- * 	};
- * }} Payload
  */
 
 // https://html.spec.whatwg.org/multipage/syntax.html#attributes-2
@@ -64,19 +44,14 @@ export const VoidElements = new Set([
 	'wbr'
 ]);
 
-/**
- * @type {Element | null}
- */
-let current_element = null;
-
-/** @returns {Payload} */
+/** @returns {import('#server').Payload} */
 function create_payload() {
 	return { out: '', head: { title: '', out: '', anchor: 0 }, anchor: 0 };
 }
 
 /**
- * @param {Payload} to_copy
- * @returns {Payload}
+ * @param {import('#server').Payload} to_copy
+ * @returns {import('#server').Payload}
  */
 export function copy_payload(to_copy) {
 	return {
@@ -87,8 +62,8 @@ export function copy_payload(to_copy) {
 
 /**
  * Assigns second payload to first
- * @param {Payload} p1
- * @param {Payload} p2
+ * @param {import('#server').Payload} p1
+ * @param {import('#server').Payload} p2
  * @returns {void}
  */
 export function assign_payload(p1, p2) {
@@ -98,59 +73,7 @@ export function assign_payload(p1, p2) {
 }
 
 /**
- * @param {Payload} payload
- * @param {string} message
- */
-function error_on_client(payload, message) {
-	message =
-		`Svelte SSR validation error:\n\n${message}\n\n` +
-		'Ensure your components render valid HTML as the browser will try to repair invalid HTML, ' +
-		'which may result in content being shifted around and will likely result in a hydration mismatch.';
-	// eslint-disable-next-line no-console
-	console.error(message);
-	payload.head.out += `<script>console.error(\`${message}\`)</script>`;
-}
-
-/**
- * @param {string} tag
- * @param {Payload} payload
- */
-export function push_element(tag, payload) {
-	if (current_element !== null && !is_tag_valid_with_parent(tag, current_element.tag)) {
-		error_on_client(payload, `<${tag}> is invalid inside <${current_element.tag}>`);
-	}
-	if (interactive_elements.has(tag)) {
-		let element = current_element;
-		while (element !== null) {
-			if (interactive_elements.has(element.tag)) {
-				error_on_client(payload, `<${tag}> is invalid inside <${element.tag}>`);
-			}
-			element = element.parent;
-		}
-	}
-	if (disallowed_paragraph_contents.includes(tag)) {
-		let element = current_element;
-		while (element !== null) {
-			if (element.tag === 'p') {
-				error_on_client(payload, `<${tag}> is invalid inside <p>`);
-			}
-			element = element.parent;
-		}
-	}
-	current_element = {
-		tag,
-		parent: current_element
-	};
-}
-
-export function pop_element() {
-	if (current_element !== null) {
-		current_element = current_element.parent;
-	}
-}
-
-/**
- * @param {Payload} payload
+ * @param {import('#server').Payload} payload
  * @param {string} tag
  * @param {() => void} attributes_fn
  * @param {() => void} children_fn
@@ -214,8 +137,8 @@ export function render(component, options) {
 }
 
 /**
- * @param {Payload} payload
- * @param {(head_payload: Payload['head']) => void} fn
+ * @param {import('#server').Payload} payload
+ * @param {(head_payload: import('#server').Payload['head']) => void} fn
  * @returns {void}
  */
 export function head(payload, fn) {
@@ -239,7 +162,7 @@ export function attr(name, value, boolean) {
 }
 
 /**
- * @param {Payload} payload
+ * @param {import('#server').Payload} payload
  * @param {boolean} is_html
  * @param {Record<string, string>} props
  * @param {() => void} component
@@ -261,65 +184,49 @@ export function css_props(payload, is_html, props, component) {
 }
 
 /**
- * @param {Record<string, unknown>[]} attrs
- * @param {boolean} lowercase_attributes
- * @param {boolean} is_html
- * @param {string} class_hash
- * @param {{ styles: Record<string, string> | null; classes: string }} [additional]
+ * @param {Record<string, unknown>} attrs
+ * @param {Record<string, string>} [classes]
+ * @param {Record<string, string>} [styles]
+ * @param {number} [flags]
  * @returns {string}
  */
-export function spread_attributes(attrs, lowercase_attributes, is_html, class_hash, additional) {
-	/** @type {Record<string, unknown>} */
-	const merged_attrs = {};
-	let key;
+export function spread_attributes(attrs, classes, styles, flags = 0) {
+	if (styles) {
+		attrs.style = attrs.style
+			? style_object_to_string(merge_styles(/** @type {string} */ (attrs.style), styles))
+			: style_object_to_string(styles);
+	}
 
-	for (let i = 0; i < attrs.length; i++) {
-		const obj = attrs[i];
-		for (key in obj) {
-			// omit functions
-			if (typeof obj[key] !== 'function') {
-				merged_attrs[key] = obj[key];
+	if (classes) {
+		const classlist = attrs.class ? [attrs.class] : [];
+
+		for (const key in classes) {
+			if (classes[key]) {
+				classlist.push(key);
 			}
 		}
-	}
 
-	const styles = additional?.styles;
-	if (styles) {
-		if ('style' in merged_attrs) {
-			merged_attrs.style = style_object_to_string(
-				merge_styles(/** @type {string} */ (merged_attrs.style), styles)
-			);
-		} else {
-			merged_attrs.style = style_object_to_string(styles);
-		}
-	}
-
-	if (class_hash) {
-		if ('class' in merged_attrs) {
-			merged_attrs.class += ` ${class_hash}`;
-		} else {
-			merged_attrs.class = class_hash;
-		}
-	}
-	const classes = additional?.classes;
-	if (classes) {
-		if ('class' in merged_attrs) {
-			merged_attrs.class += ` ${classes}`;
-		} else {
-			merged_attrs.class = classes;
-		}
+		attrs.class = classlist.join(' ');
 	}
 
 	let attr_str = '';
 	let name;
 
-	for (name in merged_attrs) {
+	const is_html = (flags & ELEMENT_IS_NAMESPACED) === 0;
+	const lowercase = (flags & ELEMENT_PRESERVE_ATTRIBUTE_CASE) === 0;
+
+	for (name in attrs) {
+		// omit functions, internal svelte properties and invalid attribute names
+		if (typeof attrs[name] === 'function') continue;
+		if (name[0] === '$' && name[1] === '$') continue; // faster than name.startsWith('$$')
 		if (INVALID_ATTR_NAME_CHAR_REGEX.test(name)) continue;
-		if (lowercase_attributes) {
+
+		if (lowercase) {
 			name = name.toLowerCase();
 		}
+
 		const is_boolean = is_html && DOMBooleanAttributes.includes(name);
-		attr_str += attr(name, merged_attrs[name], is_boolean);
+		attr_str += attr(name, attrs[name], is_boolean);
 	}
 
 	return attr_str;
@@ -354,7 +261,7 @@ export function stringify(value) {
 /** @param {Record<string, string>} style_object */
 function style_object_to_string(style_object) {
 	return Object.keys(style_object)
-		.filter(/** @param {any} key */ (key) => style_object[key])
+		.filter(/** @param {any} key */ (key) => style_object[key] != null && style_object[key] !== '')
 		.map(/** @param {any} key */ (key) => `${key}: ${escape_html(style_object[key], true)};`)
 		.join(' ');
 }
@@ -497,8 +404,8 @@ export async function value_or_fallback_async(value, fallback) {
 }
 
 /**
- * @param {Payload} payload
- * @param {void | ((payload: Payload, props: Record<string, unknown>) => void)} slot_fn
+ * @param {import('#server').Payload} payload
+ * @param {void | ((payload: import('#server').Payload, props: Record<string, unknown>) => void)} slot_fn
  * @param {Record<string, unknown>} slot_props
  * @param {null | (() => void)} fallback_fn
  * @returns {void}
@@ -621,6 +528,8 @@ export function once(get_value) {
 
 export { push, pop } from './context.js';
 
+export { push_element, pop_element } from './dev.js';
+
 export {
 	add_snippet_symbol,
 	validate_component,
@@ -630,3 +539,5 @@ export {
 } from '../shared/validate.js';
 
 export { escape_html as escape };
+
+export { default_slot } from '../client/dom/legacy/misc.js';